+++ src/actors/Enemy.cpp
@@ -1,5 +1,5 @@
 #include "Enemy.h"
 
-Enemy::Enemy(const std::filesystem::path &spriteLocation, double posX, double posY, bool isDead)
+Enemy::Enemy(const std::filesystem::path &spriteLocation, int posX, int posY, bool isDead)
     : Actor(spriteLocation, posX, posY), dead(isDead) {}
